+++ products/error_tracking/backend/hogql_queries/error_tracking_query_runner.py
@@ -1,39 +1,20 @@
-import re
 import datetime
-from dataclasses import dataclass
 from zoneinfo import ZoneInfo
-
-from django.core.exceptions import ValidationError
 
 import structlog
 
-from posthog.schema import (
-    CachedErrorTrackingQueryResponse,
-    DateRange,
-    ErrorTrackingQuery,
-    ErrorTrackingQueryResponse,
-    HogQLFilters,
-)
+from posthog.schema import CachedErrorTrackingQueryResponse, ErrorTrackingQuery, ErrorTrackingQueryResponse
 
-from posthog.hogql import ast
 from posthog.hogql.constants import LimitContext
 
 from posthog.hogql_queries.insights.paginators import HogQLHasMorePaginator
 from posthog.hogql_queries.query_runner import AnalyticsQueryRunner
-from posthog.models.filters.mixins.utils import cached_property
-from posthog.models.property.util import property_to_django_filter
 from posthog.utils import relative_date_parse
 
-from products.error_tracking.backend.api.issues import ErrorTrackingIssuePreviewSerializer
-from products.error_tracking.backend.models import ErrorTrackingIssue
+from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_v1 import ErrorTrackingQueryV1Builder
+from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_v2 import ErrorTrackingQueryV2Builder
 
 logger = structlog.get_logger(__name__)
-
-
-@dataclass
-class VolumeOptions:
-    date_range: DateRange
-    resolution: int
 
 
 class ErrorTrackingQueryRunner(AnalyticsQueryRunner[ErrorTrackingQueryResponse]):
@@ -71,882 +52,44 @@ class ErrorTrackingQueryRunner(AnalyticsQueryRunner[ErrorTrackingQueryResponse])
 
     @classmethod
     def parse_relative_date_from(cls, date: str | None) -> datetime.datetime:
-        """
-        Parses a relative date string into a datetime object.
-        This is used to convert the date range from the query into a datetime object.
-        """
         if date == "all" or date is None:
-            return datetime.datetime.now(tz=ZoneInfo("UTC")) - datetime.timedelta(days=365 * 4)  # 4 years ago
-
+            return datetime.datetime.now(tz=ZoneInfo("UTC")) - datetime.timedelta(days=365 * 4)
         return relative_date_parse(date, now=datetime.datetime.now(tz=ZoneInfo("UTC")), timezone_info=ZoneInfo("UTC"))
 
     @classmethod
     def parse_relative_date_to(cls, date: str | None) -> datetime.datetime:
-        """
-        Parses a relative date string into a datetime object.
-        This is used to convert the date range from the query into a datetime object.
-        """
         if not date:
             return datetime.datetime.now(tz=ZoneInfo("UTC"))
         if date == "all":
             raise ValueError("Invalid date range")
-
         return relative_date_parse(date, ZoneInfo("UTC"), increase=True)
 
-    def to_query(self) -> ast.SelectQuery:
-        return ast.SelectQuery(
-            select=self.select_expressions(),
-            select_from=ast.JoinExpr(table=ast.Field(chain=["events"]), alias="e"),
-            where=self.where,
-            group_by=[ast.Field(chain=["id"])],
-            order_by=[ast.OrderExpr(expr=ast.Field(chain=[self.query.orderBy]), order=self.order_direction)],
-        )
-
-    def to_query_v2(self) -> ast.SelectQuery:
-        """Single ClickHouse query joining events with system.error_tracking_issues via postgres connector.
-
-        The postgres side receives simple predicates (team_id, status), while the JOIN condition
-        (e.issue_id = issues.id) is evaluated in ClickHouse.
-
-        first_seen comes from a grouped subquery over system.error_tracking_issue_fingerprints,
-        which records the true historical first occurrence regardless of the queried date range.
-        """
-        # Subquery: min(first_seen) per issue from fingerprint history
-        fingerprints_subquery = ast.SelectQuery(
-            select=[
-                ast.Field(chain=["issue_id"]),
-                ast.Alias(
-                    alias="first_seen",
-                    expr=ast.Call(name="min", args=[ast.Field(chain=["first_seen"])]),
-                ),
-            ],
-            select_from=ast.JoinExpr(
-                table=ast.Field(chain=["system", "error_tracking_issue_fingerprints"]),
-            ),
-            group_by=[ast.Field(chain=["issue_id"])],
-        )
-
-        return ast.SelectQuery(
-            select=self.select_expressions_v2(),
-            select_from=ast.JoinExpr(
-                table=ast.Field(chain=["events"]),
-                alias="e",
-                next_join=ast.JoinExpr(
-                    join_type="INNER JOIN",
-                    table=ast.Field(chain=["system", "error_tracking_issues"]),
-                    alias="issues",
-                    constraint=ast.JoinConstraint(
-                        constraint_type="ON",
-                        expr=ast.CompareOperation(
-                            op=ast.CompareOperationOp.Eq,
-                            left=ast.Field(chain=["e", "issue_id"]),
-                            right=ast.Field(chain=["issues", "id"]),
-                        ),
-                    ),
-                    next_join=ast.JoinExpr(
-                        join_type="LEFT JOIN",
-                        table=fingerprints_subquery,
-                        alias="fp",
-                        constraint=ast.JoinConstraint(
-                            constraint_type="ON",
-                            expr=ast.CompareOperation(
-                                op=ast.CompareOperationOp.Eq,
-                                left=ast.Field(chain=["fp", "issue_id"]),
-                                right=ast.Field(chain=["issues", "id"]),
-                            ),
-                        ),
-                        next_join=ast.JoinExpr(
-                            join_type="LEFT JOIN",
-                            table=ast.Field(chain=["system", "error_tracking_issue_assignments"]),
-                            alias="assignment",
-                            constraint=ast.JoinConstraint(
-                                constraint_type="ON",
-                                expr=ast.CompareOperation(
-                                    op=ast.CompareOperationOp.Eq,
-                                    left=ast.Field(chain=["assignment", "issue_id"]),
-                                    right=ast.Field(chain=["issues", "id"]),
-                                ),
-                            ),
-                        ),
-                    ),
-                ),
-            ),
-            where=self.where_v2,
-            group_by=[
-                ast.Field(chain=["issues", "id"]),
-                ast.Field(chain=["issues", "status"]),
-                ast.Field(chain=["issues", "name"]),
-                ast.Field(chain=["issues", "description"]),
-                ast.Field(chain=["fp", "first_seen"]),
-                ast.Field(chain=["assignment", "user_id"]),
-                ast.Field(chain=["assignment", "role_id"]),
-            ],
-            order_by=[ast.OrderExpr(expr=ast.Field(chain=[self.query.orderBy]), order=self.order_direction)],
-        )
-
-    def innermost_frame_attribute(self, materialized_col):
-        return ast.Call(
-            name="argMax",
-            args=[
-                ast.TupleAccess(
-                    tuple=ast.Field(chain=["properties", materialized_col]),
-                    index=-1,
-                ),
-                ast.Field(chain=["timestamp"]),
-            ],
-        )
-
-    def select_expressions(self):
-        exprs: list[ast.Expr] = [
-            ast.Alias(alias="id", expr=ast.Field(chain=["e", "issue_id"])),
-            ast.Alias(alias="last_seen", expr=ast.Call(name="max", args=[ast.Field(chain=["timestamp"])])),
-            ast.Alias(alias="first_seen", expr=ast.Call(name="min", args=[ast.Field(chain=["timestamp"])])),
-            ast.Alias(alias="function", expr=self.innermost_frame_attribute("$exception_functions")),
-            ast.Alias(alias="source", expr=self.innermost_frame_attribute("$exception_sources")),
-        ]
-
-        if self.query.withAggregations:
-            exprs.extend(
-                [
-                    ast.Alias(
-                        alias="occurrences",
-                        expr=ast.Call(name="count", distinct=True, args=[ast.Field(chain=["uuid"])]),
-                    ),
-                    ast.Alias(
-                        alias="sessions",
-                        expr=ast.Call(
-                            name="count",
-                            distinct=True,
-                            args=[
-                                ast.Call(name="nullIf", args=[ast.Field(chain=["$session_id"]), ast.Constant(value="")])
-                            ],
-                        ),
-                    ),
-                    ast.Alias(
-                        alias="users",
-                        expr=ast.Call(
-                            name="count",
-                            distinct=True,
-                            args=[
-                                ast.Call(
-                                    name="coalesce",
-                                    args=[
-                                        ast.Call(
-                                            name="nullIf",
-                                            args=[
-                                                ast.Call(
-                                                    name="toString",
-                                                    args=[ast.Field(chain=["person_id"])],
-                                                ),
-                                                ast.Constant(value="00000000-0000-0000-0000-000000000000"),
-                                            ],
-                                        ),
-                                        ast.Field(chain=["distinct_id"]),
-                                    ],
-                                )
-                            ],
-                        ),
-                    ),
-                    ast.Alias(
-                        alias="volumeRange",
-                        expr=self.select_sparkline_array(self.date_from, self.date_to, self.query.volumeResolution),
-                    ),
-                ]
-            )
-
-        if self.query.withFirstEvent:
-            exprs.append(
-                ast.Alias(
-                    alias="first_event",
-                    expr=ast.Call(
-                        name="argMin",
-                        args=[
-                            ast.Tuple(
-                                exprs=[
-                                    ast.Field(chain=["uuid"]),
-                                    ast.Field(chain=["distinct_id"]),
-                                    ast.Field(chain=["timestamp"]),
-                                    ast.Field(chain=["properties"]),
-                                ]
-                            ),
-                            ast.Field(chain=["timestamp"]),
-                        ],
-                    ),
-                )
-            )
-
-        if self.query.withLastEvent:
-            exprs.append(
-                ast.Alias(
-                    alias="last_event",
-                    expr=ast.Call(
-                        name="argMax",
-                        args=[
-                            ast.Tuple(
-                                exprs=[
-                                    ast.Field(chain=["uuid"]),
-                                    ast.Field(chain=["distinct_id"]),
-                                    ast.Field(chain=["timestamp"]),
-                                    ast.Field(chain=["properties"]),
-                                ]
-                            ),
-                            ast.Field(chain=["timestamp"]),
-                        ],
-                    ),
-                )
-            )
-
-        exprs.append(
-            ast.Alias(
-                alias="library",
-                expr=ast.Call(
-                    name="argMax", args=[ast.Field(chain=["properties", "$lib"]), ast.Field(chain=["timestamp"])]
-                ),
-            )
-        )
-
-        return exprs
-
-    def select_expressions_v2(self) -> list[ast.Expr]:
-        """Like select_expressions but sources id/status/name/description from the joined issues table,
-        and first_seen from fingerprint history (fp.first_seen) rather than the queried event window."""
-        exprs: list[ast.Expr] = [
-            ast.Alias(alias="id", expr=ast.Field(chain=["issues", "id"])),
-            ast.Alias(alias="status", expr=ast.Field(chain=["issues", "status"])),
-            ast.Alias(alias="name", expr=ast.Field(chain=["issues", "name"])),
-            ast.Alias(alias="description", expr=ast.Field(chain=["issues", "description"])),
-            ast.Alias(alias="last_seen", expr=ast.Call(name="max", args=[ast.Field(chain=["timestamp"])])),
-            ast.Alias(alias="first_seen", expr=ast.Field(chain=["fp", "first_seen"])),
-            ast.Alias(alias="assignee_user_id", expr=ast.Field(chain=["assignment", "user_id"])),
-            ast.Alias(alias="assignee_role_id", expr=ast.Field(chain=["assignment", "role_id"])),
-            ast.Alias(alias="function", expr=self.innermost_frame_attribute("$exception_functions")),
-            ast.Alias(alias="source", expr=self.innermost_frame_attribute("$exception_sources")),
-        ]
-
-        if self.query.withAggregations:
-            exprs.extend(
-                [
-                    ast.Alias(
-                        alias="occurrences",
-                        expr=ast.Call(name="count", distinct=True, args=[ast.Field(chain=["uuid"])]),
-                    ),
-                    ast.Alias(
-                        alias="sessions",
-                        expr=ast.Call(
-                            name="count",
-                            distinct=True,
-                            args=[
-                                ast.Call(name="nullIf", args=[ast.Field(chain=["$session_id"]), ast.Constant(value="")])
-                            ],
-                        ),
-                    ),
-                    ast.Alias(
-                        alias="users",
-                        expr=ast.Call(
-                            name="count",
-                            distinct=True,
-                            args=[
-                                ast.Call(
-                                    name="coalesce",
-                                    args=[
-                                        ast.Call(
-                                            name="nullIf",
-                                            args=[
-                                                ast.Call(
-                                                    name="toString",
-                                                    args=[ast.Field(chain=["person_id"])],
-                                                ),
-                                                ast.Constant(value="00000000-0000-0000-0000-000000000000"),
-                                            ],
-                                        ),
-                                        ast.Field(chain=["distinct_id"]),
-                                    ],
-                                )
-                            ],
-                        ),
-                    ),
-                    ast.Alias(
-                        alias="volumeRange",
-                        expr=self.select_sparkline_array(self.date_from, self.date_to, self.query.volumeResolution),
-                    ),
-                ]
-            )
-
-        if self.query.withFirstEvent:
-            exprs.append(
-                ast.Alias(
-                    alias="first_event",
-                    expr=ast.Call(
-                        name="argMin",
-                        args=[
-                            ast.Tuple(
-                                exprs=[
-                                    ast.Field(chain=["uuid"]),
-                                    ast.Field(chain=["distinct_id"]),
-                                    ast.Field(chain=["timestamp"]),
-                                    ast.Field(chain=["properties"]),
-                                ]
-                            ),
-                            ast.Field(chain=["timestamp"]),
-                        ],
-                    ),
-                )
-            )
-
-        if self.query.withLastEvent:
-            exprs.append(
-                ast.Alias(
-                    alias="last_event",
-                    expr=ast.Call(
-                        name="argMax",
-                        args=[
-                            ast.Tuple(
-                                exprs=[
-                                    ast.Field(chain=["uuid"]),
-                                    ast.Field(chain=["distinct_id"]),
-                                    ast.Field(chain=["timestamp"]),
-                                    ast.Field(chain=["properties"]),
-                                ]
-                            ),
-                            ast.Field(chain=["timestamp"]),
-                        ],
-                    ),
-                )
-            )
-
-        exprs.append(
-            ast.Alias(
-                alias="library",
-                expr=ast.Call(
-                    name="argMax", args=[ast.Field(chain=["properties", "$lib"]), ast.Field(chain=["timestamp"])]
-                ),
-            )
-        )
-
-        return exprs
-
-    def select_sparkline_array(self, date_from: datetime.datetime, date_to: datetime.datetime, resolution: int):
-        """
-        This function partitions a given time range into segments (or "buckets") based on the specified resolution and then computes the number of events occurring in each segment.
-        The resolution determines the total number of segments in the time range.
-        Accordingly, the duration of each segment is dictated by the total time range and the resolution.
-
-        The equivalent SQL would look like:
-            WITH
-                toDateTime('2025-03-01 00:00:00') AS date_from,
-                toDateTime('2025-03-20 00:00:00') AS date_to,
-                10 AS resolution,
-            SELECT
-                sumForEach(
-                    arrayMap(
-                        bin ->
-                            IF(
-                                timestamp > bin AND dateDiff('seconds', bin, timestamp) < dateDiff('seconds', date_from, date_to) / resolution,
-                                1,
-                                0
-                            ), ## If we are inside the right bucket, return 1, otherwise 0
-                        arrayMap(
-                            i -> dateAdd(
-                                start_time,
-                                toIntervalSecond(i * dateDiff('seconds', date_from, date_to) / resolution)
-                            ),
-                            range(0, resolution)
-                        ) ## Generate an array of len resolution containing the start times of each segment
-                    )
-                ) AS counts
-        """
-        start_time = ast.Call(
-            name="toDateTime",
-            args=[
-                ast.Constant(value=date_from),
-            ],
-        )
-        end_time = ast.Call(
-            name="toDateTime",
-            args=[
-                ast.Constant(value=date_to),
-            ],
-        )
-        total_size = ast.Call(
-            name="dateDiff",
-            args=[
-                ast.Constant(value="seconds"),
-                start_time,
-                end_time,
-            ],
-        )
-        bin_size = ast.ArithmeticOperation(
-            op=ast.ArithmeticOperationOp.Div,
-            left=total_size,
-            right=ast.Constant(value=resolution),
-        )
-        bin_timestamps = ast.Call(
-            name="arrayMap",
-            args=[
-                ast.Lambda(
-                    args=["i"],
-                    expr=ast.Call(
-                        name="dateAdd",
-                        args=[
-                            start_time,
-                            ast.Call(
-                                name="toIntervalSecond",
-                                args=[
-                                    ast.ArithmeticOperation(
-                                        op=ast.ArithmeticOperationOp.Mult, left=ast.Field(chain=["i"]), right=bin_size
-                                    )
-                                ],
-                            ),
-                        ],
-                    ),
-                ),
-                ast.Call(
-                    name="range",
-                    args=[
-                        ast.Constant(value=0),
-                        ast.Constant(value=resolution),
-                    ],
-                ),
-            ],
-        )
-        hot_indices = ast.Call(
-            name="arrayMap",
-            args=[
-                ast.Lambda(
-                    args=["bin"],
-                    expr=ast.Call(
-                        name="if",
-                        args=[
-                            ast.And(
-                                exprs=[
-                                    ast.CompareOperation(
-                                        op=ast.CompareOperationOp.Gt,
-                                        left=ast.Field(chain=["timestamp"]),
-                                        right=ast.Field(chain=["bin"]),
-                                    ),
-                                    ast.CompareOperation(
-                                        op=ast.CompareOperationOp.LtEq,
-                                        left=ast.Call(
-                                            name="dateDiff",
-                                            args=[
-                                                ast.Constant(value="seconds"),
-                                                ast.Field(chain=["bin"]),
-                                                ast.Field(chain=["timestamp"]),
-                                            ],
-                                        ),
-                                        right=bin_size,
-                                    ),
-                                ]
-                            ),
-                            ast.Constant(value=1),
-                            ast.Constant(value=0),
-                        ],
-                    ),
-                ),
-                bin_timestamps,
-            ],
-        )
-        summed = ast.Call(
-            name="sumForEach",
-            args=[hot_indices],
-        )
-        return summed
-
-    def _base_where_exprs(self) -> list[ast.Expr]:
-        """Common WHERE expressions shared between v1 and v2 queries."""
-        exprs: list[ast.Expr] = [
-            ast.CompareOperation(
-                op=ast.CompareOperationOp.Eq,
-                left=ast.Field(chain=["event"]),
-                right=ast.Constant(value="$exception"),
-            ),
-            ast.Call(
-                name="isNotNull",
-                args=[ast.Field(chain=["e", "issue_id"])],
-            ),
-            ast.Placeholder(expr=ast.Field(chain=["filters"])),
-        ]
-
-        if self.date_from:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.GtEq,
-                    left=ast.Field(chain=["timestamp"]),
-                    right=ast.Call(
-                        name="toDateTime",
-                        args=[ast.Constant(value=self.date_from)],
-                    ),
-                )
-            )
-
-        if self.date_to:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.LtEq,
-                    left=ast.Field(chain=["timestamp"]),
-                    right=ast.Call(
-                        name="toDateTime",
-                        args=[ast.Constant(value=self.date_to)],
-                    ),
-                )
-            )
-
-        if self.query.issueId:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.Eq,
-                    left=ast.Field(chain=["e", "issue_id"]),
-                    right=ast.Constant(value=self.query.issueId),
-                )
-            )
-
-        if self.query.personId:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.Eq,
-                    left=ast.Field(chain=["person_id"]),
-                    right=ast.Constant(value=self.query.personId),
-                )
-            )
-
-        if self.query.groupKey and self.query.groupTypeIndex is not None:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.Eq,
-                    left=ast.Field(chain=[f"$group_{self.query.groupTypeIndex}"]),
-                    right=ast.Constant(value=self.query.groupKey),
-                )
-            )
-
-        if self.query.searchQuery:
-            # TODO: Refine this so it only searches the frames inside $exception_list
-            # TODO: We'd eventually need a more efficient searching strategy
-            # TODO: Add fuzzy search support
-
-            # first parse the search query to split it into words, except for quoted strings
-            # then search for each word in the exception properties
-            tokens = search_tokenizer(self.query.searchQuery)
-            and_exprs: list[ast.Expr] = []
-
-            if len(tokens) > 100:
-                raise ValidationError("Too many search tokens")
-
-            for token in tokens:
-                if not token:
-                    continue
-
-                or_exprs: list[ast.Expr] = []
-
-                props_to_search = {
-                    ("properties",): [
-                        "$exception_types",
-                        "$exception_values",
-                        "$exception_sources",
-                        "$exception_functions",
-                        "email",
-                    ],
-                    ("person", "properties"): [
-                        "email",
-                    ],
-                }
-                for chain_prefix, properties in props_to_search.items():
-                    for prop in properties:
-                        or_exprs.append(
-                            ast.CompareOperation(
-                                op=ast.CompareOperationOp.Gt,
-                                left=ast.Call(
-                                    name="position",
-                                    args=[
-                                        ast.Call(name="lower", args=[ast.Field(chain=[*chain_prefix, prop])]),
-                                        ast.Call(name="lower", args=[ast.Constant(value=token)]),
-                                    ],
-                                ),
-                                right=ast.Constant(value=0),
-                            )
-                        )
-
-                and_exprs.append(ast.Or(exprs=or_exprs))
-
-            exprs.append(ast.And(exprs=and_exprs))
-
-        return exprs
-
-    @property
-    def where(self):
-        exprs = self._base_where_exprs()
-
-        # We do this prefetching of a list of "valid" issue id's based on issue properties that aren't in
-        # CH, so that when we run the aggregation and LIMIT, we can filter out the invalid issue id's
-        # This is a hack - it'll break down if the list of valid issue id's is too long, but we do it for now
-        prefetched_ids = self.prefetch_issue_ids()
-        if prefetched_ids:
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.In,
-                    left=ast.Field(chain=["e", "issue_id"]),
-                    right=ast.Constant(value=prefetched_ids),
-                )
-            )
-
-        return ast.And(exprs=exprs)
-
-    @property
-    def where_v2(self):
-        """WHERE clause for v2: filters applied directly, no Postgres prefetch needed."""
-        exprs = self._base_where_exprs()
-
-        # Status filter — simple predicate, pushed to Postgres via the issues JOIN
-        if self.query.status and self.query.status != "all":
-            exprs.append(
-                ast.CompareOperation(
-                    op=ast.CompareOperationOp.Eq,
-                    left=ast.Field(chain=["issues", "status"]),
-                    right=ast.Constant(value=self.query.status),
-                )
-            )
-
-        # Assignee filter — pushed to Postgres via the assignment LEFT JOIN
-        if self.query.assignee:
-            if self.query.assignee.type == "user":
-                exprs.append(
-                    ast.CompareOperation(
-                        op=ast.CompareOperationOp.Eq,
-                        left=ast.Field(chain=["assignment", "user_id"]),
-                        right=ast.Constant(value=self.query.assignee.id),
-                    )
-                )
-            else:
-                exprs.append(
-                    ast.CompareOperation(
-                        op=ast.CompareOperationOp.Eq,
-                        left=ast.Field(chain=["assignment", "role_id"]),
-                        right=ast.Constant(value=str(self.query.assignee.id)),
-                    )
-                )
-
-        return ast.And(exprs=exprs)
-
     def _calculate(self):
-        query = self.to_query_v2() if self.query.useChPostgresJoin else self.to_query()
+        if self.query.useChPostgresJoin:
+            builder: ErrorTrackingQueryV1Builder | ErrorTrackingQueryV2Builder = ErrorTrackingQueryV2Builder(
+                self.query, self.date_from, self.date_to
+            )
+        else:
+            builder = ErrorTrackingQueryV1Builder(self.query, self.team, self.date_from, self.date_to)
 
         with self.timings.measure("error_tracking_query_hogql_execute"):
             query_result = self.paginator.execute_hogql_query(
-                query=query,
+                query=builder.build_query(),
                 team=self.team,
                 query_type="ErrorTrackingQuery",
                 timings=self.timings,
                 modifiers=self.modifiers,
                 limit_context=self.limit_context,
-                filters=HogQLFilters(
-                    filterTestAccounts=self.query.filterTestAccounts,
-                    properties=self.hogql_properties,
-                ),
+                filters=builder.hogql_filters(),
             )
 
         columns: list[str] = query_result.columns or []
-        results = (
-            self.results_v2(columns, query_result.results)
-            if self.query.useChPostgresJoin
-            else self.results(columns, query_result.results)
-        )
 
         return ErrorTrackingQueryResponse(
             columns=columns,
-            results=results,
+            results=builder.process_results(columns, query_result.results),
             timings=query_result.timings,
             hogql=query_result.hogql,
             modifiers=self.modifiers,
             **self.paginator.response_params(),
         )
-
-    def results(self, columns: list[str], query_results: list):
-        results = []
-        mapped_results = [dict(zip(columns, value)) for value in query_results]
-        issue_ids = [result["id"] for result in mapped_results]
-
-        with self.timings.measure("issue_fetching_execute"):
-            issues = self.error_tracking_issues(issue_ids)
-
-        with self.timings.measure("issue_resolution"):
-            for result_dict in mapped_results:
-                issue = issues.get(str(result_dict["id"]))
-
-                if issue:
-                    results.append(
-                        issue
-                        | {
-                            "last_seen": result_dict.get("last_seen"),
-                            "library": result_dict.get("library"),
-                            "function": result_dict.get("function"),
-                            "source": result_dict.get("source"),
-                            "first_event": (
-                                self.extract_event(result_dict.get("first_event"))
-                                if self.query.withFirstEvent
-                                else None
-                            ),
-                            "last_event": (
-                                self.extract_event(result_dict.get("last_event")) if self.query.withLastEvent else None
-                            ),
-                            "aggregations": (
-                                self.extract_aggregations(result_dict) if self.query.withAggregations else None
-                            ),
-                        }
-                    )
-
-        return results
-
-    def results_v2(self, columns: list[str], query_results: list):
-        """Process results from the v2 query (single CH+postgres join, no separate Postgres round-trip)."""
-        results = []
-        for row in query_results:
-            result_dict = dict(zip(columns, row))
-            results.append(
-                {
-                    "id": str(result_dict["id"]),
-                    "status": result_dict.get("status"),
-                    "name": result_dict.get("name"),
-                    "description": result_dict.get("description"),
-                    "first_seen": result_dict.get("first_seen"),
-                    "assignee": self._extract_assignee_v2(result_dict),
-                    "last_seen": result_dict.get("last_seen"),
-                    "library": result_dict.get("library"),
-                    "function": result_dict.get("function"),
-                    "source": result_dict.get("source"),
-                    "first_event": (
-                        self.extract_event(result_dict.get("first_event")) if self.query.withFirstEvent else None
-                    ),
-                    "last_event": (
-                        self.extract_event(result_dict.get("last_event")) if self.query.withLastEvent else None
-                    ),
-                    "aggregations": self.extract_aggregations(result_dict) if self.query.withAggregations else None,
-                }
-            )
-        return results
-
-    def _extract_assignee_v2(self, result_dict: dict) -> dict | None:
-        user_id = result_dict.get("assignee_user_id")
-        role_id = result_dict.get("assignee_role_id")
-        if user_id:
-            return {"id": user_id, "type": "user"}
-        if role_id:
-            return {"id": str(role_id), "type": "role"}
-        return None
-
-    def extract_event(self, event_tuple):
-        if event_tuple is None:
-            return None
-        else:
-            properties = event_tuple[3]
-
-            return {
-                "uuid": str(event_tuple[0]),
-                "distinct_id": str(event_tuple[1]),
-                "timestamp": str(event_tuple[2]),
-                "properties": properties,
-            }
-
-    def get_volume_buckets(self) -> list[datetime.datetime]:
-        if self.query.volumeResolution == 0:
-            return []
-        total_ms = (self.date_to - self.date_from).total_seconds() * 1000
-        bin_size = int(total_ms / self.query.volumeResolution)
-        return [
-            self.date_from + datetime.timedelta(milliseconds=i * bin_size) for i in range(self.query.volumeResolution)
-        ]
-
-    def extract_aggregations(self, result):
-        # TODO: Remove unused volumeRange. (keeping it for now because of cached values)
-        aggregations = {f: result[f] for f in ("occurrences", "sessions", "users", "volumeRange")}
-        histogram_bins = self.get_volume_buckets()
-        aggregations["volume_buckets"] = [
-            {"label": bin.isoformat(), "value": aggregations["volumeRange"][i] if aggregations["volumeRange"] else None}
-            for i, bin in enumerate(histogram_bins)
-        ]
-        return aggregations
-
-    @property
-    def order_direction(self):
-        if self.query.orderDirection:
-            return self.query.orderDirection.value
-
-        return "ASC" if self.query.orderBy == "first_seen" else "DESC"
-
-    def error_tracking_issues(self, ids):
-        status = self.query.status
-        queryset = (
-            ErrorTrackingIssue.objects.with_first_seen().select_related("assignment").filter(team=self.team, id__in=ids)
-        )
-
-        if self.query.issueId:
-            queryset = queryset.filter(id=self.query.issueId)
-        elif status and not status == "all":
-            queryset = queryset.filter(status=status)
-
-        if self.query.assignee:
-            queryset = (
-                queryset.filter(assignment__user_id=self.query.assignee.id)
-                if self.query.assignee.type == "user"
-                else queryset.filter(assignment__role_id=self.query.assignee.id)
-            )
-
-        for filter in self.issue_properties:
-            queryset = property_to_django_filter(queryset, filter)
-
-        serializer = ErrorTrackingIssuePreviewSerializer(queryset, many=True)
-        return {issue["id"]: issue for issue in serializer.data}
-
-    def prefetch_issue_ids(self) -> list[str]:
-        # We hit postgres to get a list of "valid" issue id's based on issue properties that aren't in
-        # CH, but that we want to filter the returned results by. This is a hack - it'll break down if
-        # the list of valid issue id's is too long, but we do it for now, until we can get issue properties
-        # into CH
-
-        use_prefetched = False
-        if self.query.issueId:
-            # If we have an issueId, we should just use that
-            return [self.query.issueId]
-
-        queryset = ErrorTrackingIssue.objects.with_first_seen().select_related("assignment").filter(team=self.team)
-
-        if self.query.status and self.query.status not in ["all", "active"]:
-            use_prefetched = True
-            queryset = queryset.filter(status=self.query.status)
-
-        if self.query.assignee:
-            use_prefetched = True
-            queryset = (
-                queryset.filter(assignment__user_id=self.query.assignee.id)
-                if self.query.assignee.type == "user"
-                else queryset.filter(assignment__role_id=str(self.query.assignee.id))
-            )
-
-        for filter in self.issue_properties:
-            queryset = property_to_django_filter(queryset, filter)
-
-        if not use_prefetched:
-            return []
-
-        return [str(issue["id"]) for issue in queryset.values("id")]
-
-    @cached_property
-    def issue_properties(self):
-        return [value for value in self.properties if "error_tracking_issue" == value.type]
-
-    @cached_property
-    def hogql_properties(self):
-        return [value for value in self.properties if "error_tracking_issue" != value.type]
-
-    @cached_property
-    def properties(self):
-        return self.query.filterGroup.values[0].values if self.query.filterGroup else []
-
-
-def search_tokenizer(query: str) -> list[str]:
-    # parse the search query to split it into words, except for quoted strings. Strip quotes from quoted strings.
-    # Example: 'This is a "quoted string" and this is \'another one\' with some words'
-    # Output: ['This', 'is', 'a', 'quoted string', 'and', 'this', 'is', 'another one', 'with', 'some', 'words']
-    # This doesn't handle nested quotes, and some complex edge cases, but we don't really need that for now.
-    # If requirements do change, consider using a proper parser like `pyparsing`
-    pattern = r'"[^"]*"|\'[^\']*\'|\S+'
-    tokens = re.findall(pattern, query)
-    return [token.strip("'\"") for token in tokens]
